agents/command_agent/requirements.txt
fastapi
uvicorn
groq
pydantic
python-dotenv